+++ src/components/HealthWidget/HealthWidget.tsx
@@ -8,9 +8,9 @@ import HealthInfoDisplay from './HealthInfoDisplay';
 import HealthInfoEdit from './HealthInfoEdit';
 
 // Util Imports
-import { getPetById, getMedicationsByPet } from '../../utils/api';
+import { getPetById, getMedicationsByPet } from '../../api';
 
-import type { Pet } from '../../utils/api/services/pets/types';
+import type { Pet } from '../../api/services/pets/types';
 
 interface Props {
   petId: number;
